+++ modules/aurora_core/include/AuroraTestModule.h
@@ -23,6 +23,8 @@ private:
 
 	void RunViewTests();
 
+	void RunMovementSystemTests();
+
 	void RunSystemManagerTests();
 
 	void RunECSTests();