+++ modules/aurora_core/include/ecs/systems/MovementSystem.h
@@ -5,12 +5,19 @@
  *
  * File:
  *      MovementSystem.h
+ *
+ * Description:
+ *      Updates TransformComponent using VelocityComponent.
  ******************************************************************************/
 
 #include "ecs/systems/ISystem.h"
 
 class MovementSystem : public ISystem {
 public:
+	// -------------------------------------------------------------------------
+	// ISystem
+	// -------------------------------------------------------------------------
+
 	void Update(
 			Registry &registry,
 			float deltaTime) override;