+++ modules/aurora_core/src/ecs/systems/MovementSystem.cpp
@@ -1,7 +1,8 @@
 #include "ecs/systems/MovementSystem.h"
 
-#include "AuroraLogger.h"
 #include "ecs/Registry.h"
+#include "ecs/components/TransformComponent.h"
+#include "ecs/components/VelocityComponent.h"
 
 // -----------------------------------------------------------------------------
 // Update
@@ -10,6 +11,23 @@
 void MovementSystem::Update(
 		Registry &registry,
 		float deltaTime) {
-	AuroraLogger::Info(
-			"MovementSystem Update");
+	auto view = registry.CreateView<TransformComponent>();
+
+	for (EntityID entityId : view) {
+		Entity entity(entityId);
+
+		if (!registry.HasComponent<VelocityComponent>(entity)) {
+			continue;
+		}
+
+		auto &transform =
+				registry.GetComponent<TransformComponent>(entity);
+
+		auto &velocity =
+				registry.GetComponent<VelocityComponent>(entity);
+
+		transform.x += velocity.x * deltaTime;
+		transform.y += velocity.y * deltaTime;
+		transform.z += velocity.z * deltaTime;
+	}
 }